+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## [Unreleased]
+### Fixed
+- List Records - Fix filter on belongs to with name different from fieldName.
 
 ## RELEASE 3.3.5 - 2019-09-26
 ### Changed